feat: add class DataPool
